implemented delete tours restriction to only adminand tour-guide only

+++ starter/controllers/tour-controller.js
@@ -1,5 +1,6 @@
 const Tour = require('./../models/tourModel');
 const ApiFeature = require('./../utils/apiFeature');
+const AppError = require(`${__dirname}/../utils/globalError`);
 
  var month = [
    'January',
@@ -149,9 +150,16 @@ exports.createTour = async (req, res) => {
   }
 };
 
-exports.getTour = async (req, res) => {
+exports.getTour = async (req, res , next) => {
   try {
-    const tour = await Tour.findById(req.params.id);
+
+     const tour = await Tour.findById(req.params.id);
+     if (!tour) {
+       return next(new AppError('Tour not found ', 404));
+     }
+
+    // const tour = await Tour.findById(req.params.id);
+    console.log(tour)
     res.status(200).json({
       status: true,
       message: 'success',
@@ -166,9 +174,16 @@ exports.getTour = async (req, res) => {
   }
 };
 
-exports.updateTour = async (req, res) => {
+exports.updateTour = async (req, res , next) => {
   try {
-    const tour = await Tour.findByIdAndUpdate(req.params.id, req.body, {
+    const tour = await Tour.findById(req.params.id)
+
+      if (! tour) {
+        return next(new AppError('Tour not found ', 404));
+      }
+
+
+    const updatedtour = await Tour.findByIdAndUpdate(tour._id, req.body, {
       new: true,
       runValidators: true,
     }); //returns new updated tour
@@ -176,7 +191,7 @@ exports.updateTour = async (req, res) => {
     res.status(200).json({
       status: true,
       message: 'success',
-      data: { tour },
+      data: { tour:updatedtour },
     });
   } catch (error) {
     console.log('ERROR ' + error.message);
@@ -187,8 +202,14 @@ exports.updateTour = async (req, res) => {
   }
 };
 
-exports.deleteTour = async (req, res) => {
+exports.deleteTour = async (req, res , next) => {
   try {
+    console.log(req.params)
+   const tour = await Tour.findById(req.params.id);
+    if (! tour) {
+      return next(new AppError('Tour not found ', 404));
+    }
+
     await Tour.findByIdAndDelete(req.params.id); //returns new updated tour
 
     res.status(204).json({
